+++ src/containerapp/azext_containerapp/commands.py
@@ -42,7 +42,7 @@ def load_command_table(self, args):
     with self.command_group('containerapp env http-route-config', is_preview=True) as g:
         g.custom_show_command('show', 'show_httprouteconfig')
         g.custom_command('list', 'list_httprouteconfigs')
-        g.custom_command('create', 'update_httprouteconfig', exception_handler=ex_handler_factory())
+        g.custom_command('create', 'create_httprouteconfig', exception_handler=ex_handler_factory())
         g.custom_command('update', 'update_httprouteconfig', exception_handler=ex_handler_factory())
         g.custom_command('delete', 'delete_httprouteconfig', confirmation=True, exception_handler=ex_handler_factory())